Add new JS file to separate functinality
